requirements.txt
Django==4.0.6
psycopg2-binary
gunicorn
python-dotenv